funcionalidad de accesso a los roles de administrador

+++ spring-ecommerce/src/main/java/com/cursos/ecommerce/spring_ecommerce/models/DetalleOrden.java
@@ -17,7 +17,7 @@ public class DetalleOrden {
     private Long id;
 
     private String nombre;
-    private double cantidad;
+    private Integer cantidad;
     private double precio;
     private double total;
